Major improvements to the Game Library. Added ImGui as the main GUI library for the Studio.
Game library gets heavy improvements to make it more straightforward to make a tight and efficient pipeline. Added new object: Game Package which is used to submit all the game codes to the engine. Some components were moved to the runtime directory to give easy access and better classification. ImGui is used as the default GUI library for the studio and will be compiled as a static library file and linked to the Dynamik Studio executable.