Basics of the new Game Library is in place. The complete architecture is in Documentation/README/GameLibrary.md file. License is added to the engine.
The basic game library components are in place (Player Object, Level Component, Game Library and more). The low level functionalities are yet to be made along with proper game and level management. The complete architcture is in Documentation/README/GameLibrary.md file. Apache 2.0 license is added to the engine and the repository will be made public.