Major speed improvements. Added new mesh loading mechanism. Minor game library improments.
Removed unnecessary wrapping of pointers using the POINTER<TYPE> class which would not make any difference in the release and distribution modes but has a significant performance hit in debug mode. Added utility functions to support some of the functionalities proveded by the POINTER class. Mesh data loading is now much faster as it does not allocate vertexes when loading data. It now allocates a single buffer using the vertex layout and requested data are added to the buffer while loading data. Minor improvements to the game library provides functions to perform some tasks easily and added a new game component: Environment Map.