A folder was renamed. Added a developer snapshot and added very basic support for animation control. A new data type is being added.
Core/Object folder was renamed to Core/Objects in Source/Runtime directory.

A developer snapshot was added to the repository and a minor bug was fixed at the rendering API's frame buffer resize handle and context termination.
Added basic support for the animation controller.

A new data type: Quaternion is being added to the engine.

The Game Library documentation was updated.