Basic animation support added.
The engine has basic support for animation rendering. It now has containers to store animation data but the pipeline is not yet made.

Added new functions to the math library to work with quaternions and the vector class is expected to be updated as a minor memory bug was identified (in SIMD result storing).