Moved some of the core components to the core module. Added more features to the animation module.
The core components of the engine (mesh, physics body, ...) were previously in the component system module. Now it is in the core module as it is a core component of the engine.

Animation module now has a animated mesh component which the engine uses to render the animation (skinned mesh + animation (bone mesh + key frames)). A basic animated mesh importer is in place. Improvements and tests are yet to be made.