A folder was renamed. Improvements to the animation module.
The folder "Core/Objects/Resource/" was renamed to "Core/Objects/Resources/". All the components and types in the Primitive file was moved to their respective files.

The dynamik animation structure will contain all the transforms of a single animation based on its duration and ticks-per-second. Additionally all the node data will be stored for later calculations.